create test for JPA delete beer not found

+++ src/main/java/com/pjsdev/spring6restmvc/services/BeerServiceJPA.java
@@ -54,8 +54,12 @@ public class BeerServiceJPA implements BeerService {
     }
 
     @Override
-    public void deleteBeerById(UUID id) {
-        beerRepository.deleteById(id);
+    public Boolean deleteBeerById(UUID id) {
+        if (beerRepository.existsById(id)) {
+            beerRepository.deleteById(id);
+            return true;
+        }
+        return false;
     }
 
     @Override
